fixed bug with adding new element to recycleView

+++ app/src/main/java/com/example/epamandroid/backend/IWebService.kt
@@ -10,5 +10,5 @@ interface IWebService<T> {
         callback: ICallback<List<T>>
     )
 
-    fun removeEntity(id: Long?)
+    fun removeEntity(id: Int?)
 }
