+++ app/src/main/java/com/example/epamandroid/backend/entities/StudentModel.kt
@@ -1,3 +1,3 @@
 package com.example.epamandroid.backend.entities
 
-class StudentModel(val id: Long, val name: String, val hwCount: Int,  val isStudent: Boolean)
+class StudentModel(val id: Int, val name: String, val hwCount: Int,  val isStudent: Boolean)
